+++ v3/crates/engine/benches/execute.rs
@@ -59,6 +59,15 @@ pub fn bench_execute(
     )
     .unwrap();
 
+    let validate_non_null_graphql_variables = if resolved_metadata
+        .runtime_flags
+        .contains(metadata_resolve::flags::ResolvedRuntimeFlag::ValidateNonNullGraphqlVariables)
+    {
+        gql::validation::NonNullGraphqlVariablesValidation::Validate
+    } else {
+        gql::validation::NonNullGraphqlVariablesValidation::DoNotValidate
+    };
+
     let gds = GDS {
         metadata: Arc::new(resolved_metadata.clone()),
     };
@@ -127,6 +136,7 @@ pub fn bench_execute(
                     },
                     schema,
                     request,
+                    validate_non_null_graphql_variables,
                 )
                 .unwrap();
             });
@@ -139,6 +149,7 @@ pub fn bench_execute(
         },
         &schema,
         &request,
+        validate_non_null_graphql_variables,
     )
     .unwrap();